Began to implement item interface and item type enum. Began to mock text window displays in Display Drafts.html
