+++ DandD/Character.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using DandD.Enums;
+using DandD.Interfaces;
 
 namespace DandD
 {
