--- DandD/ClassTypes.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace DandD
-{
-        public enum ClassTypes
-        {
-            Barbarian,
-            Bard,
-            Cleric,
-            Druid,
-            Fighter,
-            Wizard,
-            Monk,
-            Paladin,
-            Ranger,
-            Sorcerer,
-            Rogue,           
-        }
-    
-}
