+++ DandD/DisplayDrafts.html
@@ -1,0 +1,76 @@
+﻿<!DOCTYPE html>
+<html>
+<head>
+	<title>Different Display Drafts</title>
+	<meta charset="utf-8" />
+	<link href="https://fonts.googleapis.com/css?family=Cinzel:400,700,900" rel="stylesheet">
+	<link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet" />
+	<style>
+		body, p, h1, h2, h3, h4, h5, h6 { margin: 0; }
+		h1 { color: #000; font-size: 40px; font-weight: bold; margin: 15px; }
+		h3 { font-weight: bold; margin-left: 10px; }
+		[class*="-window"] { margin: 0 10px 30px 10px; padding: 30px; background-color: #ffeeca; }
+		p { font-family: 'Cinzel'; font-weight: 400; font-size: 40px; line-height: 1.2; text-justify: inter-word; }
+			p:first-letter { font-weight: 700; font-size: 50px; margin-left: 30px; }
+		.boss-window { border: 5px inset #bb0000; background: linear-gradient(135deg, #ffeeca 90%,#bb0000 100%); }
+		.boss-text:first-letter { color: #bb0000; }
+		.dark-window { border: 5px inset #330070; background: linear-gradient(135deg, #ffeeca 90%,#330070 100%); }
+		.dark-text:first-letter { color: #330070; }
+		.forest-window { border: 5px inset #005500; background: linear-gradient(135deg, #ffeeca 90%,#005500 100%); }
+		.forest-text:first-letter { color: #005500; }
+		.indoor-window { border: 5px inset #5e4502; background: linear-gradient(135deg, #ffeeca 90%,#5e4502 100%); }
+		.indoor-text:first-letter { color: #5e4502; }
+		.light-window { border: 5px inset #0062b5; background: linear-gradient(135deg, #ffeeca 90%,#0062b5 100%); }
+		.light-text:first-letter { color: #0062b5; }
+		.outdoor-window { border: 5px inset #ffd80f; background: linear-gradient(135deg, #ffeeca 90%,#ffd80f 100%); }
+		.outdoor-text:first-letter { color: #ffd80f; }
+	</style>
+</head>
+<body>
+	<section class="text-box">
+		<h1>Text Window Diplays</h1>
+		<div class="row">
+			<div class="col-sm-6">
+				<h3>Boss Battle Text Window</h3>
+				<div class="boss-window">
+					<p class="boss-text">
+						Flames engulf your wooden shield before your trembling hands can reach it. Why you bring a wood shield to a dragon boss battle?
+					</p>
+				</div>
+				<h3>Dark Places Text Window</h3>
+				<div class="dark-window">
+					<p class="dark-text">
+						Don't look down. It's best not to know what you stepped in, or where it could have come out of.
+					</p>
+				</div>
+				<h3>Open Area Text Window</h3>
+				<div class="light-window">
+					<p class="light-text">
+						What a glorious view of the vast wheat fields of spring. Not a foe in sight. Except for the nasty whole wheat demon.
+					</p>
+				</div>
+			</div>
+			<div class="col-sm-6">
+				<h3>Forest Text Window</h3>
+				<div class="forest-window">
+					<p class="forest-text">
+						This particular forest exhibits an unmistakable aroma of magic and elf cheese. It is home to the ancient Elves of the Dark Cheese.
+					</p>
+				</div>
+				<h3>Indoor Text Window</h3>
+				<div class="indoor-window">
+					<p class="indoor-text">
+						"You won't need your weapon here, you're amongst friends!", the barkeep said before you opened his throat with your greataxe.
+					</p>
+				</div>
+				<h3>Outdoor Text Window</h3>
+				<div class="outdoor-window">
+					<p class="outdoor-text">
+						There is a crowded mass of townsfolk in the market today. How many can you unjustifiably slaughter?
+					</p>
+				</div>
+			</div>
+		</div>
+	</section>
+</body>
+</html>
