+++ DandD/Enums/AlignmentTypes.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace DandD.Enums
+{
+    public enum AlignmentTypes
+    {
+        Lawful_Good,
+        Nuetral_Good,
+        Chaotic_Good,
+        Lawful_Nuetral,
+        Nuetral_Nuetral,
+        Chaotic_Nuetral,
+        Lawful_Evil,
+        Nuetral_Evil,
+        Chaotic_Evil,
+    }
+}
