+++ DandD/Enums/ClassTypes.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace DandD.Enums
+{
+    public enum ClassTypes
+    {
+        Barbarian,
+        Bard,
+        Cleric,
+        Druid,
+        Fighter,
+        Wizard,
+        Monk,
+        Paladin,
+        Ranger,
+        Sorcerer,
+        Rogue,           
+    }
+}
