+++ DandD/Enums/ItemTypes.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace DandD.Enums
+{
+	public enum ItemTypes
+	{
+		Consumeable,		
+		Wearable,
+		Weapon,
+		Armor,
+		Utility,
+		Literary
+	}
+}
