+++ DandD/Enums/RaceTypes.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace DandD.Enums
+{
+    public enum RaceTypes
+    {        
+        Dwarf,
+        Elf,
+        Gnome,
+        Half_Elf,
+        Halk_Orc,
+        Halfling,
+        Human     
+    }
+}
