--- DandD/RaceTypes.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace DandD
-{
-    public enum RaceTypes
-    {        
-        Dwarf,
-        Elf,
-        Gnome,
-        Half_Elf,
-        Halk_Orc,
-        Halfling,
-        Human     
-    }
-}
